Trap/SummonedHouseNpc ctor byte<->sbyte: (sbyte) cast NPC_DATA.GetNpcTemplate().GetLevel() arg3 to SummonedObject base ctor (sbyte level; faithful Java byte=sbyte); clears 2 CS1503 (897->895)

+++ dotnetConversion/src/Aion.GameServer/Model/GameObjects/SummonedHouseNpc.cs
@@ -12,7 +12,7 @@ namespace Aion.GameServer.Model.GameObjects;
 public class SummonedHouseNpc : SummonedObject<House.House>
 {
     public SummonedHouseNpc(NpcController controller, SpawnTemplate spawnTemplate, House.House house)
-        : base(controller, spawnTemplate, DataManager.NPC_DATA.GetNpcTemplate(spawnTemplate.GetNpcId()).GetLevel(), house)
+        : base(controller, spawnTemplate, (sbyte)DataManager.NPC_DATA.GetNpcTemplate(spawnTemplate.GetNpcId()).GetLevel(), house)
     {
         string masterName = house.GetOwnerName();
         SetMasterName(masterName == null ? "" : masterName);
